% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{arrange_split_txt}
\alias{arrange_split_txt}
\title{Wrapper function of converting text into arranged text grob.}
\usage{
arrange_split_txt(..., width = NULL, silent = TRUE)
}
\arguments{
\item{...}{text.}

\item{width}{grid::unit}

\item{silent}{Logical. TRUE: no massage, FALSE: shows massage.}
}
\value{
textGrob
}
\description{
Wrapper function of converting text into arranged text grob.
}
\examples{
txt_title    <- "This is a title."
txt_ITEM_1   <- "ITEM text is larger than item text."
txt_item_2   <- "This is a item text, too."
txt_BODY_1   <- "BODY text is larger than body text."
txt_body_2   <- "This is a body text, too."
arranged_txt_1 <- arrange_split_txt(txt_title, txt_ITEM_1, txt_item_2)
# purrr::map(arranged_txt_1, grid::grid.draw)

}
