% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_textgrob.R
\name{arrange_txt}
\alias{arrange_txt}
\title{Wrapper function of converting text into arranged text grob.}
\usage{
arrange_txt(..., hjust = 1, widths = grid::unit(841, "mm"), silent = TRUE)
}
\arguments{
\item{...}{Text object(s). Multiple object can be input.
Object name of text: foo_keyword or foo_keyword_bar
examples: txt_title, txt_body_1, proj_ITEM_big
keyword: "tiny", "script", "footnote", "caption",
"body", "BODY", "item", "ITEM",
"thank", "author", "affil",
"subt", "kw" or "title"}

\item{hjust}{A numeric 0-1. 1: left", 0: "right"}

\item{widths}{grid::unit}

\item{silent}{Logical. TRUE: no massage, FALSE: shows massage.}
}
\value{
\preformatted{   textGrob
}
}
\description{
Wrapper function of converting text into arranged text grob.
}
\examples{
txt_title    <- "This is a title."
txt_ITEM_1   <- "ITEM text is larger than item text."
txt_item_2   <- "This is a item text, too."
txt_BODY_1   <- "BODY text is larger than body text."
txt_body_2   <- "This is a body text, too."
arranged_txt_1 <- arrange_txt(txt_title, txt_ITEM_1, txt_item_2)
grid::grid.draw(arranged_txt_1)
arranged_txt_2 <- arrange_txt(txt_BODY_1, txt_body_2, hjust=1)
grid::grid.draw(arranged_txt_2)

}
