% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_grob.R
\name{as_stg}
\alias{as_stg}
\title{Wrapper function to generate split textGrob for grid::textGrob.}
\usage{
as_stg(text, width = NULL, font_size, use = NULL, ...)
}
\arguments{
\item{text}{A string.
label in grid::textGrob().}

\item{font_size}{A numeric of font size.
fontsize in grid::textGrob().}

\item{use}{A string.
name in grid::textGrob().}

\item{just}{A string.
"left", "right", "centre", "center", "bottom", and "top".
See grid::textGrob() in detail.}
}
\description{
Wrapper function to generate split textGrob for grid::textGrob.
}
\seealso{
grid::textGrob()
}
