% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_grob.R
\name{as_tg}
\alias{as_tg}
\title{Wrapper function to generate textGrob for grid::textGrob.}
\usage{
as_tg(text, font_size, use, hjust, ...)
}
\arguments{
\item{text}{A string.
label in grid::textGrob().}

\item{font_size}{A numeric of font size.
fontsize in grid::textGrob().}

\item{use}{A string.
name in grid::textGrob().}

\item{hjust}{A numeric 0-1. 1: left", 0: "right"}

\item{...}{Some more arguments}
}
\description{
Wrapper function to generate textGrob for grid::textGrob.
}
\seealso{
grid::textGrob()
}
