% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_class}
\alias{check_class}
\alias{check_text}
\alias{check_numeric}
\title{Check class of object.}
\usage{
check_text(x)

check_numeric(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
TRUE or FALSE and warning message.
}
\description{
Check class of object.
}
\examples{
x <- 1
y <- "text"
check_text(x)
check_text(y)
check_numeric(x)
check_numeric(y)

}
