% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grobs.R
\name{combine_iamge_grobs}
\alias{combine_iamge_grobs}
\alias{appose_image_grobs}
\alias{stack_image_grobs}
\title{Appose or stack two grobs or more in a line (row or col).}
\usage{
appose_image_grobs(
  ...,
  width = NULL,
  height = NULL,
  grow = TRUE,
  unify = "height",
  space = grid::unit(0, "mm"),
  gp = grid::gpar(),
  name = NULL
)

stack_image_grobs(
  ...,
  width = NULL,
  height = NULL,
  grow = TRUE,
  unify = "width",
  space = grid::unit(0, "mm"),
  gp = grid::gpar(),
  name = NULL
)
}
\arguments{
\item{...}{grob.}

\item{width, height}{A grid unit.}

\item{grow}{Do not work yet!
A string. "height", "width", or "none"
Images will be grown to fill the width or height.
Work ONLY when both of width and height are set.}

\item{unify}{A string.}

\item{space}{A grid unit. Space among grobs.}

\item{gp}{gpar() object.}

\item{name}{A string. grob name}
}
\value{
\preformatted{        combined (apposed or stacked) grob in a line.
}
}
\description{
Appose or stack two grobs or more in a line (row or col).
}
\examples{
library(grid)
small <- rectGrob(width=unit(1, "cm"), height=unit(1, "cm"), gp=gpar(fill="black"))
tall  <- rectGrob(width=unit(1, "cm"), height=unit(2, "cm"), gp=gpar(fill="red"))
wide  <- rectGrob(width=unit(2, "cm"), height=unit(1, "cm"), gp=gpar(fill="blue"))
large <- rectGrob(width=unit(2, "cm"), height=unit(2, "cm"), gp=gpar(fill="white"))
combined_grobs <-
  (small \%oo\% tall \%oo\% wide \%oo\% large) \%8\% # original
  appose_image_grobs(small, tall, wide, large, space=grid::unit(1, "mm"))
grid.draw(combined_grobs)

}
