% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{drawDetails.splitText}
\alias{drawDetails.splitText}
\title{split string draw}
\usage{
\method{drawDetails}{splitText}(x, recording)
}
\arguments{
\item{x}{A text grob}

\item{recording}{A logical value indicating whether a grob is being added to
the display list or redrawn from the display list.}
}
\description{
split string draw
}
