% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grobs.R
\name{frame_place_grobs}
\alias{frame_place_grobs}
\title{Frame and place grobs in line (row or col)}
\usage{
frame_place_grobs(
  grobs,
  layout,
  widths,
  heights,
  space,
  by_row = TRUE,
<<<<<<< HEAD
=======
  gp = grid::gpar(),
>>>>>>> develop
  name = NULL
)
}
\arguments{
\item{grobs}{grobs to be combined}

\item{layout}{grid layout.}

\item{widths}{grid unit.}

\item{heights}{grid unit.}

\item{space}{A grid unit. Space among grobs.}

\item{by_row}{A logical. TRUE: line in a row. FALSE: line in a col.}

<<<<<<< HEAD
=======
\item{gp}{gpar().}

>>>>>>> develop
\item{name}{A string. Name of combined grob.}
}
\value{
\preformatted{      combined grobs by layout.
}
}
\description{
This function is used in appose_image_grobs() and stack_image_grobs().
}
