% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_slide.R
\name{gen_slide}
\alias{gen_slide}
\title{Generate slide.}
\usage{
gen_slide(
  title,
  fig,
  text,
  gp_title,
  gp_text,
  paper = "",
  width = unit(1, "npc"),
  height = unit(1, "npc"),
  layout = NULL,
  name = NULL,
  ...
)
}
\arguments{
\item{title}{A text.}

\item{fig}{ggplot object.}

\item{text}{A body text}

\item{gp_title}{gpar of title}

\item{gp_text}{gpar of text}

\item{paper}{Slide size. default: "a4".}

\item{width}{Slide width.}

\item{height}{Slide height.}

\item{layout}{Layout of slide.}

\item{name}{A name of grob.}
}
\value{
Grob
}
\description{
Generate slide.
}
\examples{
  # title
title <- "THIS IS TITLE WORDS"
  # gp_title
fontsize <- 40
gp_title <- grid::gpar(fontsize=fontsize)
  # figure
data("mpg", package="ggplot2")
fig <- 
  ggplot2::ggplot(mpg) + 
  ggplot2::geom_point(ggplot2::aes(x = .data[["displ"]], y = .data[["hwy"]])) +
  ggplot2::theme_bw()
fig  # draw ggplot object
  # text
t_1 <- "This is a sample text. "
t_2 <- "When longer than width, "
t_3 <- "text will be split and add breaks. "
t_4 <- "This text is separated into 5 objects, "
t_5 <- "because of R document specification."
text <- stringr::str_c(t_1, t_2, t_3, t_4, t_5)
  # gp_text
fontsize <- 20
lineheight <- 1.1
gp_text <- grid::gpar(fontsize=fontsize, lineheight=lineheight)
  # generate slide and draw
slide <- gen_slide(title, fig, text, gp_title, gp_text)
grid::grid.draw(slide)

}
