% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_grob.R
\name{get_font_size}
\alias{get_font_size}
\title{Get font size from font_size_list with name or use.}
\usage{
get_font_size(base = NULL, name = NULL, use = NULL, silent = TRUE, text = "")
}
\arguments{
\item{base}{A numeric of base font size:
8, 10, 12, 14, 18, 22, 26, 32, or 38)
If base=NULL, base=26 will be set in get_font_size().}

\item{name}{A string of font size name.
"tiny", "scriptsize", "footnotesize", "small",
"normalsize", "large", "Large", "LARGE",
"huge", "Huge", or "HUGE"}

\item{use}{A string of font size name for use.
"tiny", "script", "footnote", "caption",
"body", "BODY", "item", "ITEM", "thank",
"author", "affil", "subt", "kw", or "title"
thank: acknowledgement
affil: affiliation
subt: subtitle
kw: keyword}

\item{silent}{Logical. TRUE: no massage, FALSE: shows massage.}

\item{text}{A string, but this argument will be omitted.}
}
\value{
A numeric of font size.
}
\description{
Get font size from font_size_list with name or use.
}
\examples{
get_font_size(base=NULL, name="large", use=NULL)
get_font_size(base=NULL, name="large", use="title")
get_font_size(base=NULL, name=NULL,    use="ITEM")
get_font_size(base=22,   name=NULL,    use="kw")

}
\seealso{
font_size_list
}
