% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggposter.R
\docType{package}
\name{ggposter-package}
\alias{ggposter}
\alias{ggposter-package}
\title{ggposter: build A1 conference posters from ggplot2}
\description{
ggposter declares a poster as a title band plus columns of rounded,
tab-headed cards (text, tables, figures, or photo strips), assembled with
\link[patchwork:patchwork-package]{patchwork::patchwork} and rendered at true size with embedded fonts. See
\code{\link[=poster]{poster()}} to build one from an R list or YAML file, and the package
vignette (\code{vignette("ggposter")}) for a full walkthrough.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matutosi.github.io/ggposter/}
  \item \url{https://github.com/matutosi/ggposter}
  \item Report bugs at \url{https://github.com/matutosi/ggposter/issues}
}

}
\author{
\strong{Maintainer}: Toshikazu Matsumura \email{matutosi@gmail.com}

}
\keyword{internal}
