% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inprogress.R
\name{in_progress}
\alias{in_progress}
\title{in progress}
\description{
in progress
}
\examples{
# This is only memo for author. 
# D:/matu/work/ToDo/ggposter/tools/splitstring.R
# D:/matu/work/ToDo/ggposter/tools/test.R
# D:/matu/work/ToDo/ggposter/tools/utils2.R
}
