% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grobs.R
\name{reverse_ratio}
\alias{reverse_ratio}
\title{Compute rev_ratio of length to max length}
\usage{
reverse_ratio(lengths)
}
\arguments{
\item{lengths}{grid units.}
}
\value{
\preformatted{      Numerics.
}
}
\description{
This function is used in appose_image_grobs() and stack_image_grobs().
}
\examples{
len <- grid::unit(c(1 ,2, 4), "npc")
reverse_ratio(len)

}
