% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_string.R
\name{split_text}
\alias{split_text}
\alias{splitString}
\alias{splitTextGrob}
\alias{drawDetails.splitText}
\title{Split text to fit viewpeort width}
\usage{
splitString(text)

splitTextGrob(text, ...)

drawDetails.splitText(x, recording)
}
\arguments{
\item{text}{A string.}

\item{...}{See grid::grob()}

\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\value{
textGrob
}
\description{
This function is a little bit modification of splitString in "R Graphics"
Chapter 6:  Developing New Graphics Functions and Objects
https://www.stat.auckland.ac.nz/~paul/RGraphics/interactgrid-calcdraw.R
}
\examples{
# grid::grid.newpage()
t_1 <- "This is a sample text. "
t_2 <- "When longer than viewpeort width, "
t_3 <- "text will be split and add a break. "
t_4 <- "This text is separated into 5 objects, "
t_5 <- "because of R document specification."
text <- stringr::str_c(t_1, t_2, t_3, t_4, t_5)
fontsize <- 16
stg <- splitTextGrob(text, name="test", gp=grid::gpar(fontsize=fontsize))
grid::grid.draw(stg)

}
