% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_grob.R
\name{split_text_grob}
\alias{split_text_grob}
\alias{splitString}
\title{Split text to fit viewpeort width}
\usage{
split_text_grob(text, width = NULL, gp = NULL, name = NULL)

splitString(text)
}
\arguments{
\item{text}{A string.}

\item{width}{grid::unit()}

\item{gp}{grid::gpar()}

\item{name}{A string.}
}
\value{
splitString(): split text by "\\n". split_text_grob: textGrob.
}
\description{
This function is a little bit modification of splitString in "R Graphics"
Chapter 6: Developing New Graphics Functions and Objects
https://www.stat.auckland.ac.nz/~paul/RGraphics/interactgrid-calcdraw.R
}
\examples{
# grid::grid.newpage()
library(grid)
t_1 <- "This is a sample text. "
t_2 <- "When longer than viewpeort width, "
t_3 <- "text will be split and add a break. "
t_4 <- "This text is separated into 5 objects, "
t_5 <- "because of R document specification."
text <- stringr::str_c(t_1, t_2, t_3, t_4, t_5)
# Split text in the context of current grid.gpar()
splitString(text)
# Generate split textGrob on condisions of arguments.
fontsize <- 40
stg <- split_text_grob(text, gp=gpar(fontsize=40, font=2))
grid.draw(stg)

}
