% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrob.R
\name{tg2layout}
\alias{tg2layout}
\title{Layout}
\usage{
tg2layout(tg, width = NULL, height = NULL, row_margin = 1.5)
}
\arguments{
\item{tg}{A text grob}

\item{width}{A numeric.}

\item{height}{A numeric.}

\item{row_margin}{A numeric.}
}
\value{
Lauyout grob
}
\description{
Layout
}
