% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrob.R
\name{txt2tibble}
\alias{txt2tibble}
\title{Convert text into tibble.}
\usage{
txt2tibble(...)
}
\arguments{
\item{...}{Text object(s). Multiple object can be input.
Object name of text: foo_keyword or foo_keyword_bar
examples: txt_title, txt_body_1, proj_ITEM_big
keyword: "tiny", "script", "footnote", "caption",
"body", "BODY", "item", "ITEM",
"thank", "author", "affil",
"subt", "kw" or "title"}
}
\value{
Tibble with "text" and "use" colnames.
Values of "text" are character of value to be input.
Values of "use" are characters of value name to be input.
}
\description{
Convert text into tibble.
}
\examples{
txt_title <- "This is a title."
txt_ITEM_1 <- "ITEM text is larger than item text."
txt_item_2 <- "This is a item text, too."
txt_BODY_1 <- "BODY text is larger than body text."
txt_body_1 <- "This is a body text, too."
txt2tibble(txt_title, txt_ITEM_1, txt_item_2, txt_BODY_1, txt_body_1)

}
\seealso{
get_font_size(), arrange_txt(),
}
