% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{as_stg}
\alias{as_stg}
\title{Wrapper function to generate split textGrob for grid::textGrob.}
\usage{
as_stg(text, width = NULL, font_size, use = NULL, ...)
}
\arguments{
\item{text}{A string.
label in grid::textGrob().}

\item{width}{grid::unit}

\item{font_size}{A numeric of font size.
fontsize in grid::textGrob().}

\item{use}{A string.
name in grid::textGrob().}

\item{...}{Others.}
}
\description{
Wrapper function to generate split textGrob for grid::textGrob.
}
\seealso{
grid::textGrob()
}
