% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrob.R
\name{as_tg}
\alias{as_tg}
\title{Wrapper function for grid::textGrob.}
\usage{
as_tg(text, font_size, use, just)
}
\arguments{
\item{text}{A string.
label in grid::textGrob().}

\item{font_size}{A numeric of font size.
fontsize in grid::textGrob().}

\item{use}{A string.
name in grid::textGrob().}

\item{just}{A string.
"left", "right", "centre", "center", "bottom", and "top".
See grid::textGrob() in detail.}
}
\description{
Wrapper function for grid::textGrob.
}
\seealso{
grid::textGrob()
}
