% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_grob.R
\name{as_tg}
\alias{as_tg}
\title{Wrapper function to generate textGrob for grid::textGrob.}
\usage{
as_tg(text, x, y, hjust, vjust, font_size, use, ...)
}
\arguments{
\item{text}{A string.
label in grid::textGrob().}

\item{x, y}{A numeric. Specify x- or y-value.}

\item{hjust, vjust}{A numeric 0-1. 0: left, 1: right.}

\item{font_size}{A numeric of font size.
fontsize in grid::textGrob().}

\item{use}{A string. Name in grid::textGrob().}

\item{...}{Some more arguments}
}
\description{
Wrapper function to generate textGrob for grid::textGrob.
}
\seealso{
grid::textGrob()
}
