% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrob.R
\name{as_tg}
\alias{as_tg}
\title{Convert text to text grob.}
\usage{
as_tg(text, font_size = NULL, base = NULL, name = NULL, use = NULL)
}
\arguments{
\item{text}{A string.}

\item{font_size}{A numeric of font size.}

\item{base}{A numeric of base font size:
8, 10, 12, 14, 18, 22, 26, 32, or 38)
See font_size_list.
If base=NULL, base=26 will be set in get_font_size().}

\item{name}{A string of font size name.
"tiny", "scriptsize", "footnotesize", "small",
"normalsize", "large", "Large", "LARGE",
"huge", "Huge", or "HUGE"}

\item{use}{A string of font size name for use.
"tiny", "script", "footnote", "caption",
"body", "BODY", "item", "ITEM", "thank",
"author", "affil", "subt", "kw", or "title"
thank: acknowledgement
affil: affiliation
subt: subtitle
kw: keyword}
}
\description{
Convert text to text grob.
}
