% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grobs.R
\name{combine_grobs}
\alias{combine_grobs}
\alias{appose_grob}
\alias{stack_grob}
\alias{appose_grobs}
\alias{stack_grobs}
\alias{\%oo\%}
\alias{\%8\%}
\title{Appose or stack two grobs or more in a line (row or col).}
\usage{
appose_grob(x, y)

stack_grob(x, y)

appose_grobs(...)

stack_grobs(...)

x \%oo\% y

x \%8\% y
}
\arguments{
\item{x, y, ...}{grob}
}
\value{
apposed or stack grob in a line
}
\description{
appose_grob() and tack_grob() conbine two grobs,
while appose_grobs() and stack_grobs() can conbine more.
\verb{\%oo\%} and \verb{\%8\%} are shortcuts of appose_grob() and tack_grob().
}
\examples{
x <- grid::textGrob("foo_bar")
y <- grid::textGrob("xxx_yyy")
z <- grid::textGrob("hogehoge")
ag <- appose_grob(x, y)
sg <- stack_grob(x, y)
ags <- appose_grobs(x, y, z)
sgs <- stack_grobs(x, y, z)
grid::grid.draw(ag)
grid::grid.newpage()
grid::grid.draw(sg)
grid::grid.newpage()
grid::grid.draw(ags)
grid::grid.newpage()
grid::grid.draw(sgs)
grid::grid.newpage()
grid::grid.draw(y \%oo\% z)
grid::grid.newpage()
grid::grid.draw(y \%8\% z)
grid::grid.newpage()

}
