% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grobs.R
\name{combine_grobs}
\alias{combine_grobs}
\alias{appose_grob}
\alias{stack_grob}
\alias{\%oo\%}
\alias{\%8\%}
\alias{appose_grobs}
\alias{stack_grobs}
\alias{appose_grobs_conv}
\alias{stack_grobs_conv}
\title{Appose or stack two grobs or more in a line (row or col).}
\usage{
appose_grob(
  gx,
  gy,
  space = grid::unit(0, "mm"),
  gp = grid::gpar(),
  name = NULL
)

stack_grob(gx, gy, space = grid::unit(0, "mm"), gp = grid::gpar(), name = NULL)

gx \%oo\% gy

gx \%8\% gy

appose_grobs(..., space = grid::unit(0, "mm"), gp = grid::gpar(), name = NULL)

stack_grobs(..., space = grid::unit(0, "mm"), gp = grid::gpar(), name = NULL)

appose_grobs_conv(
  ...,
  space = grid::unit(0, "mm"),
  gp = grid::gpar(),
  name = NULL
)

stack_grobs_conv(
  ...,
  space = grid::unit(0, "mm"),
  gp = grid::gpar(),
  name = NULL
)
}
\arguments{
\item{gx, gy, ...}{grob.}

\item{space}{A grid unit. Space among grobs.}

\item{gp}{gpar().}

\item{name}{A string. grob name}
}
\value{
\preformatted{     combined (apposed or stacked) grob in a line.
}
}
\description{
appose_grob() and tack_grob() combine two grobs,
while appose_grobs() and stack_grobs() can combine more.
\verb{\%oo\%} and \verb{\%8\%} are shortcuts of appose_grob() and tack_grob().
}
\examples{
library(grid)
x <- grid::textGrob("foo_bar", hjust=1, vjust=0)
x2 <- grid::textGrob("foo_bar", x=0, y=1, hjust=1, vjust=1)
y <- gridtext::textbox_grob(
     "textbox_grob generate wrapped text in a box.",
     width=grid::unit(30, "mm"), box_gp=grid::gpar(col="black"))
z <- gridtext::richtext_grob(
     "richtext_grob can NOT wrap text",
     box_gp=grid::gpar(col="black"))
ag <- appose_grob(x, y, name="apppsed_grob")
sg <- stack_grob(x, y)
ags <- appose_grobs(x, y, z)
sgs <- stack_grobs(x, y, z)
grid::grid.draw(ag)
grid::grid.newpage()
grid::grid.draw(sg)
grid::grid.newpage()
grid::grid.draw(ags)
grid::grid.newpage()
grid::grid.draw(sgs)
grid::grid.newpage()
grid::grid.draw(y \%oo\% z)
grid::grid.newpage()
grid::grid.draw(y \%8\% z)
grid::grid.newpage()
grid::grid.draw(x \%oo\% y \%8\% z)
grid::grid.newpage()
grid::grid.draw(x \%oo\% (y \%8\% z))
grid::grid.newpage()
grid::grid.draw(x2 \%oo\% (y \%8\% z))

# NOTE: do no work in gridtext package when use "npc" in grid::unit
# library(gridtext)
# a <- gridtext::textbox_grob(
#      "textbox_grob generate wrapped text in a box.",
#       width=grid::unit(0.5, "npc"), box_gp=grid::gpar(col="black"))
# b <- gridtext::richtext_grob(
#      "richtext_grob can NOT wrap text",
#      box_gp=grid::gpar(col="black"))
# grid::grid.draw(a \%oo\% b)
# grid::grid.draw(a \%8\% b)

}
