% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grobs.R
\name{frame_place_grobs}
\alias{frame_place_grobs}
\title{Frame and place grobs}
\usage{
frame_place_grobs(grobs, layout, widths, heights, space, name = NULL)
}
\arguments{
\item{grobs}{grobs to be combined}

\item{layout}{grid layout.}

\item{widths}{grid unit.}

\item{heights}{grid unit.}

\item{name}{A string. Name of combined grob.}
}
\value{
\preformatted{      combined grobs by layout.
}
}
\description{
This function is used in appose_image_grobs() and stack_image_grobs().
}
