% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{split_text}
\alias{split_text}
\alias{split_string}
\alias{splitTextGrob}
\alias{split_text_grob}
\title{Split text to fit viewpeort width}
\usage{
split_string(text, width = grid::unit(1, "npc"))

splitTextGrob(text, ...)

split_text_grob(text, width = NULL, gp = NULL, name = NULL, ...)
}
\arguments{
\item{text}{A string.}

\item{width}{grid::unit}

\item{...}{See grid::grob()}

\item{gp}{gpar().}

\item{name}{A string. grob name.}
}
\value{
textGrob
}
\description{
This function is a little bit modificated vertion of splitString in "R Graphics"
Chapter 6: Developing New Graphics Functions and Objects
https://www.stat.auckland.ac.nz/~paul/RGraphics/interactgrid-calcdraw.R

Called when drawDetails are run.

Called when drawDetails are run.
}
\examples{
# grid::grid.newpage()
library(grid)
t_1 <- "This is a sample text. "
t_2 <- "When longer than viewpeort width, "
t_3 <- "text will be split and add a break. "
t_4 <- "This text is separated into 5 objects, "
t_5 <- "because of R document specification."
text <- stringr::str_c(t_1, t_2, t_3, t_4, t_5)
fontsize <- 40
stg <- splitTextGrob(text, name="test", gp=grid::gpar(fontsize=fontsize))
grid.draw(stg)
grid::grid.gedit("test", gp=grid::gpar(fontsize=40))

}
