% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrob.R
\name{tg2layout}
\alias{tg2layout}
\title{Create layout to fit height and width with text grob.}
\usage{
tg2layout(tg, width = grid::unit(841, "mm"), height = NULL, row_margin = 1.5)
}
\arguments{
\item{tg}{A text grob}

\item{width}{grid::unit}

\item{height}{A numeric. Default is null to fit text grob.}

\item{row_margin}{A numeric. Default is 1.5. This makes rows 1.5 heights of text grobs.}
}
\value{
Lauyout grob
}
\description{
Create layout to fit height and width with text grob.
}
