% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrob.R
\name{tg_arrange}
\alias{tg_arrange}
\title{Title.}
\usage{
tg_arrange(tg, layout, name = NULL)
}
\arguments{
\item{tg}{A text grob}

\item{layout}{A Grid layout, or NULL.
This can be used to initialise the frame with a number of
rows and columns, with initial widths and heights, etc.}

\item{name}{A character identifier of frame name.}
}
\description{
Title.
}
